+++ NINA.Plugin.TargetScheduler/Sequencer/TargetSchedulerFlats.cs
@@ -84,10 +84,6 @@ namespace NINA.Plugin.TargetScheduler.Sequencer {
                 await CloseCover(progress, token);
                 await ToggleLight(true, progress, token);
 
-                imageSaveMediator.BeforeImageSaved += BeforeImageSaved;
-                imageSaveMediator.BeforeFinalizeImageSaved += BeforeFinalizeImageSaved;
-                imageSaveMediator.ImageSaved += ImageSaved;
-
                 // Get the unique set of target ids over all neededFlats
                 List<int> targetIds = new List<int>();
                 foreach (LightSession neededFlat in neededFlats) {
@@ -145,13 +141,6 @@ namespace NINA.Plugin.TargetScheduler.Sequencer {
                 CompletedFlatSets = 0;
                 Iterations = 0;
                 CompletedIterations = 0;
-
-                // Wait before we stop monitoring for image saves - give the last one a chance to come through
-                Thread.Sleep(3000);
-
-                imageSaveMediator.BeforeImageSaved -= BeforeImageSaved;
-                imageSaveMediator.BeforeFinalizeImageSaved -= BeforeFinalizeImageSaved;
-                imageSaveMediator.ImageSaved -= ImageSaved;
             }
 
             return;